requirements.txt
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
requests==2.31.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pymysql==1.1.0
python-dotenv==1.0.0
streamlit-authenticator==0.1.5
streamlit-option-menu==0.3.6
bcrypt==4.0.1
python-dateutil==2.8.2
openpyxl==3.1.2
PyPDF2==3.0.1
